requirements.txt
anthropic[bedrock,vertex]>=0.40.0
pillow>=10.0.0
PyAutoGUI>=0.9.54
PyYAML>=6.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
google-genai>=1.0.0
